+++ byzantine-tests/src/it/scala/co/topl/tetra/it/util/package.scala
@@ -1,12 +1,12 @@
 package co.topl.tetra.it
 
-import co.topl.algebras.{ToplGenusRpc, ToplRpc}
+import co.topl.algebras.{ToplGenusRpc, NodeRpc}
 import com.spotify.docker.client.DockerClient
 import scala.language.implicitConversions
 
 package object util {
 
-  implicit def rpcToRpcApi[F[_]](rpc: ToplRpc[F, fs2.Stream[F, *]]): NodeRpcApi[F] =
+  implicit def rpcToRpcApi[F[_]](rpc: NodeRpc[F, fs2.Stream[F, *]]): NodeRpcApi[F] =
     new NodeRpcApi(rpc)
 
   implicit def rpcToGenusRpcApi[F[_]](rpc: ToplGenusRpc[F]): GenusRpcApi[F] =
