+++ node/src/main/resources/application.conf
@@ -113,16 +113,8 @@ bifrost {
   }
 }
 
-// Serving gRPC requires explicitly enabling HTTP2
-akka.http.server.preview.enable-http2 = on
-
 genus {
   enable = true
-  rpc-host = "0.0.0.0"
-  rpc-port = 9091
-  rpc-node-host = "localhost"
-  rpc-node-port = 9084
-  rpc-node-tls = false
 
   // The _base_ directory in which genus_db data is stored
   // optionally if OrientDB Studio was previously installed required, the genus_db storage should be a child of ¨database¨ folder
