+++ node/src/main/scala/co/topl/genus/GenusGrpc.scala
@@ -1,17 +1,19 @@
 package co.topl.genus
 
-import cats.{Eval, Now}
 import cats.effect.kernel.Async
 import cats.effect.kernel.Resource
 import cats.implicits._
+import cats.Eval
+import cats.Now
 import co.topl.algebras.ToplGenusRpc
 import co.topl.genus.services._
-import co.topl.genusLibrary.algebras.{BlockFetcherAlgebra, TransactionFetcherAlgebra, VertexFetcherAlgebra}
+import co.topl.genusLibrary.algebras.BlockFetcherAlgebra
+import co.topl.genusLibrary.algebras.TransactionFetcherAlgebra
+import co.topl.genusLibrary.algebras.VertexFetcherAlgebra
 import fs2.grpc.syntax.all._
-import io.grpc.{Metadata, Server}
-import io.grpc.netty.shaded.io.grpc.netty.{NettyChannelBuilder, NettyServerBuilder}
-import io.grpc.protobuf.services.ProtoReflectionService
-import java.net.InetSocketAddress
+import io.grpc.Metadata
+import io.grpc.ServerServiceDefinition
+import io.grpc.netty.shaded.io.grpc.netty.NettyChannelBuilder
 
 object GenusGrpc {
 
@@ -56,34 +58,16 @@ object GenusGrpc {
 
   object Server {
 
-    def serve[F[_]: Async](
-      host:               String,
-      port:               Int,
+    def services[F[_]: Async](
       blockFetcher:       BlockFetcherAlgebra[F],
       transactionFetcher: TransactionFetcherAlgebra[F],
       vertexFetcher:      VertexFetcherAlgebra[F]
-    ): Resource[F, Server] =
-      for {
-        blockService <- BlockServiceFs2Grpc.bindServiceResource(
-          new GrpcBlockService(blockFetcher)
-        )
-        transactionService <- TransactionServiceFs2Grpc.bindServiceResource(
-          new GrpcTransactionService(transactionFetcher)
-        )
-        networkMetricsService <- NetworkMetricsServiceFs2Grpc.bindServiceResource(
-          new GrpcNetworkMetricsService(vertexFetcher)
-        )
-
-        server <-
-          NettyServerBuilder
-            .forAddress(new InetSocketAddress(host, port))
-            .addService(blockService)
-            .addService(transactionService)
-            .addService(networkMetricsService)
-            .addService(ProtoReflectionService.newInstance())
-            .resource[F]
-            .evalMap(server => Async[F].delay(server.start()))
-      } yield server
+    ): Resource[F, List[ServerServiceDefinition]] =
+      List(
+        BlockServiceFs2Grpc.bindServiceResource(new GrpcBlockService(blockFetcher)),
+        TransactionServiceFs2Grpc.bindServiceResource(new GrpcTransactionService(transactionFetcher)),
+        NetworkMetricsServiceFs2Grpc.bindServiceResource(new GrpcNetworkMetricsService(vertexFetcher))
+      ).sequence
 
   }
 }
