--- node/src/main/scala/co/topl/genus/GenusServer.scala
@@ -1,101 +1,0 @@
-package co.topl.genus
-
-import cats.data.{EitherT, OptionT}
-import cats.effect._
-import cats.effect.implicits._
-import cats.implicits._
-import co.topl.algebras.{SynchronizationTraversalStep, SynchronizationTraversalSteps}
-import co.topl.codecs.bytes.tetra.instances.blockHeaderAsBlockHeaderOps
-import co.topl.genus.services.BlockData
-import co.topl.genusLibrary.interpreter._
-import co.topl.genusLibrary.model.GE
-import co.topl.genusLibrary.orientDb.OrientDBFactory
-import co.topl.genusLibrary.orientDb.OrientThread
-import co.topl.grpc.ToplGrpc
-import co.topl.node.ApplicationConfig
-import co.topl.typeclasses.implicits.showBlockId
-import org.typelevel.log4cats.Logger
-import org.typelevel.log4cats.slf4j.Slf4jLogger
-import scala.concurrent.duration._
-import scala.jdk.CollectionConverters._
-
-object GenusServer {
-
-  def make[F[_]: Async](conf: ApplicationConfig.Genus): Resource[F, Unit] =
-    if (conf.enable) for {
-      implicit0(logger: Logger[F]) <- Resource.pure(Slf4jLogger.getLoggerFromName[F]("Genus"))
-      // A dedicated single thread executor in which all OrientDB calls are expected to run
-      implicit0(orientThread: OrientThread[F]) <- OrientThread.create[F]
-      orientdb                                 <- OrientDBFactory.make[F](conf.orientDbDirectory, conf.orientDbPassword)
-
-      dbTx <- Resource
-        .eval(Async[F].delay(orientdb.getTx))
-        .evalTap(db => orientThread.delay(db.makeActive()))
-      dbNoTx <- Resource
-        .eval(Async[F].delay(orientdb.getNoTx))
-        .evalTap(db => orientThread.delay(db.makeActive()))
-
-      rpcInterpreter   <- ToplGrpc.Client.make[F](conf.rpcNodeHost, conf.rpcNodePort, conf.rpcNodeTls)
-      nodeBlockFetcher <- NodeBlockFetcher.make(rpcInterpreter)
-
-      vertexFetcher      <- GraphVertexFetcher.make[F](dbNoTx)
-      blockFetcher       <- GraphBlockFetcher.make(vertexFetcher)
-      graphBlockUpdater  <- GraphBlockUpdater.make[F](dbTx, blockFetcher, nodeBlockFetcher)
-      transactionFetcher <- GraphTransactionFetcher.make(vertexFetcher)
-
-      _ <- GenusGrpc.Server
-        .serve(conf.rpcHost, conf.rpcPort, blockFetcher, transactionFetcher, vertexFetcher)
-        .evalTap(grpcServer =>
-          Logger[F].info(s"RPC Server bound at ${grpcServer.getListenSockets.asScala.toList.mkString(",")}")
-        )
-
-      // Delay replication by 20 seconds to allow Node RPC to launch
-      processorStream: fs2.Stream[F, Unit] =
-        for {
-          _ <- fs2.Stream.sleep[F](20.seconds)
-          nodeLatestHeight <- fs2.Stream.eval(
-            OptionT(nodeBlockFetcher.fetchHeight()).getOrRaise(new IllegalStateException("Unknown node height"))
-          )
-          graphCurrentHeight <- fs2.Stream.eval(
-            OptionT(
-              blockFetcher
-                .fetchCanonicalHead()
-                .rethrow
-            ).fold(0L)(_.height)
-          )
-          _ <- fs2.Stream.eval(
-            Logger[F].info(s"Historical data start=${graphCurrentHeight + 1}, end=${nodeLatestHeight}")
-          )
-          // Historical + live data streams
-          _ <- fs2.Stream
-            .force[F, BlockData](
-              nodeBlockFetcher.fetch(startHeight = graphCurrentHeight + 1, endHeight = nodeLatestHeight + 3)
-            )
-            .evalTap(blockData => Logger[F].info(s"Inserting block data ${blockData.header.id.show}"))
-            .evalMap(graphBlockUpdater.insert)
-            .rethrow ++
-          fs2.Stream
-            .force[F, SynchronizationTraversalStep](
-              rpcInterpreter.synchronizationTraversal()
-            )
-            .evalMap {
-              case SynchronizationTraversalSteps.Applied(blockId) =>
-                EitherT(nodeBlockFetcher.fetch(blockId))
-                  .semiflatTap(blockData => Logger[F].info(s"Inserting block ${blockData.header.id.show}"))
-                  .flatMapF(graphBlockUpdater.insert)
-                  .value
-              case SynchronizationTraversalSteps.Unapplied(blockId) =>
-                EitherT(nodeBlockFetcher.fetch(blockId))
-                  .semiflatTap(blockData => Logger[F].info(s"Deleting block ${blockData.header.id.show}"))
-                  .flatMapF(graphBlockUpdater.remove)
-                  .value
-            }
-            .rethrow
-            .recover(_ => ().asRight[GE])
-
-        } yield ()
-      _ <- processorStream.compile.drain.background
-    } yield ()
-    else Resource.unit[F]
-
-}
